requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
scipy>=1.6.0
scikit-image>=0.18.0
tqdm>=4.60.0
tensorboard>=2.5.0
albumentations>=1.0.0
imgaug>=0.4.0
matplotlib>=3.3.0
